HOTFIX-5353 ORC-1078: Row group end offset doesn't accommodate all the blocks
To calculate the row group end offset based on the buffer size. Current implementation assumes stretching the slop by a constant factor of 2 to accommodate the compression block but thats not always the case when smaller buffer sizes are used.

When compression size is smaller than 2048 and two blocks are not enough for all the values left in last row group, we might end up seeking outside of range.

Unit test added TestRecordReaderImpl#testSmallCompressionSizeOrc

Change-Id: Id840f03a40fabd8f3a79aa342afda3248568e0dd