CDPD-28921: Use the common CDH version of JUnit
ORC has so far pulled in Junit v4.11 as a dependency. This had to be
upgraded to mitigate CVE-2020-15250.

The optimal solution is to harmonize the JUnit dependency to the common
JUnit version specified for the whole of CDH, which is currently 4.13.3.
Using the common CDH version will outsource the versioning to CDH, so
that ORC does not have to manage it separately.

Change-Id: Ia041e31f4ad5886cf691d8920b53311c611e60d0